+++ cider/src/flatten/flat_ir/wires/mod.rs
@@ -1,4 +1,4 @@
-/// A module containing the assignment and group definitions
-pub mod core;
 /// A module containing the assignment guard definitions
 pub mod guards;
+/// A module containing the assignment and group definitions
+pub mod structures;
