+++ cider/src/flatten/primitives/utils.rs
@@ -88,7 +88,9 @@ macro_rules! get_params {
 
 pub(crate) use get_params;
 
-use crate::flatten::{flat_ir::base::PortValue, structures::thread::ThreadIdx};
+use crate::flatten::{
+    flat_ir::indexes::PortValue, structures::thread::ThreadIdx,
+};
 
 pub fn infer_thread_id<'a, I: Iterator<Item = &'a PortValue>>(
     iter: I,
